+++ iotdb-0.10/src/main/java/cn/edu/tsinghua/iotdb/benchmark/iotdb010/IoTDB.java
@@ -5,35 +5,25 @@ import cn.edu.tsinghua.iotdb.benchmark.conf.Config;
 import cn.edu.tsinghua.iotdb.benchmark.conf.ConfigDescriptor;
 import cn.edu.tsinghua.iotdb.benchmark.conf.Constants;
 import cn.edu.tsinghua.iotdb.benchmark.measurement.Status;
-import cn.edu.tsinghua.iotdb.benchmark.tsdb.DBUtil;
 import cn.edu.tsinghua.iotdb.benchmark.tsdb.IDatabase;
 import cn.edu.tsinghua.iotdb.benchmark.tsdb.TsdbException;
 import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Batch;
 import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Record;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.AggRangeQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.AggRangeValueQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.AggValueQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.GroupByQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.LatestPointQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.PreciseQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.RangeQuery;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.ValueRangeQuery;
+import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.*;
 import cn.edu.tsinghua.iotdb.benchmark.workload.schema.DeviceSchema;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.sql.*;
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class IoTDB implements IDatabase {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(IoTDB.class);
-  private static final Config config = ConfigDescriptor.getInstance().getConfig();
+  private static Config config = ConfigDescriptor.getInstance().getConfig();
 
   private static final String CREATE_SERIES_SQL =
       "CREATE TIMESERIES %s WITH DATATYPE=%s,ENCODING=%s,COMPRESSOR=%s";
@@ -62,7 +52,7 @@ public class IoTDB implements IDatabase {
   }
 
   @Override
-  public void cleanup() {
+  public void cleanup() throws TsdbException {
     // currently no implementation
   }
 
@@ -108,7 +98,7 @@ public class IoTDB implements IDatabase {
         for (DeviceSchema deviceSchema : schemaList) {
           int sensorIndex = 0;
           for (String sensor : deviceSchema.getSensors()) {
-            String dataType = DBUtil.getDataType(sensorIndex);
+            String dataType = getNextDataType(sensorIndex);
             String createSeriesSql = String.format(CREATE_SERIES_SQL,
                 Constants.ROOT_SERIES_NAME
                     + "." + deviceSchema.getGroup()
@@ -136,7 +126,63 @@ public class IoTDB implements IDatabase {
     }
   }
 
+  String getNextDataType(int sensorIndex) {
+    List<Double> proportion = resolveDataTypeProportion();
+    double[] p = new double[6 + 1];
+    p[0] = 0.0;
+    // split [0,1] to n regions, each region corresponds to a data type whose proportion
+    // is the region range size.
+    for (int i = 1; i <= 6; i++) {
+      p[i] = p[i - 1] + proportion.get(i - 1);
+    }
+    double sensorPosition = sensorIndex * 1.0 / config.getSENSOR_NUMBER();
+    int i;
+    for (i = 1; i <= 6; i++) {
+      if (sensorPosition >= p[i - 1] && sensorPosition < p[i]) {
+        break;
+      }
+    }
+    switch (i) {
+      case 1:
+        return "BOOLEAN";
+      case 2:
+        return "INT32";
+      case 3:
+        return "INT64";
+      case 4:
+        return "FLOAT";
+      case 5:
+        return "DOUBLE";
+      case 6:
+        return "TEXT";
+      default:
+        LOGGER.error("Unsupported data type {}, use default data type: TEXT.", i);
+        return "TEXT";
+    }
+  }
 
+  List<Double> resolveDataTypeProportion() {
+    List<Double> proportion = new ArrayList<>();
+    String[] split = config.getINSERT_DATATYPE_PROPORTION().split(":");
+    if (split.length != 6) {
+      LOGGER.error("INSERT_DATATYPE_PROPORTION error, please check this parameter.");
+    }
+    double[] proportions = new double[6];
+    double sum = 0;
+    for (int i = 0; i < split.length; i++) {
+      proportions[i] = Double.parseDouble(split[i]);
+      sum += proportions[i];
+    }
+    for (int i = 0; i < split.length; i++) {
+      if (sum != 0) {
+        proportion.add(proportions[i] / sum);
+      } else {
+        proportion.add(0.0);
+        LOGGER.error("The sum of INSERT_DATATYPE_PROPORTION is zero!");
+      }
+    }
+    return proportion;
+  }
 
   String getEncodingType(String dataType) {
     switch (dataType) {
@@ -256,7 +302,7 @@ public class IoTDB implements IDatabase {
   }
 
   /**
-   * SELECT last s_76 FROM root.group_3.d_31
+   * SELECT max_time(s_76) FROM root.group_3.d_31
    */
   @Override
   public Status latestPointQuery(LatestPointQuery latestPointQuery) {
@@ -310,7 +356,7 @@ public class IoTDB implements IDatabase {
     builder.append(timestamp);
     int sensorIndex = 0;
     for (String value : values) {
-      switch (DBUtil.getDataType(sensorIndex)) {
+      switch (getNextDataType(sensorIndex)) {
         case "BOOLEAN":
           boolean tempBoolean = (Double.parseDouble(value) > 500);
           builder.append(",").append(tempBoolean);
